Remove LookupTableNameValidator validator
