--- src/lookup-table/utils/validations/error.ts
@@ -1,8 +1,0 @@
-class ValidationErrors extends Error {
-  constructor(message: string[]) {
-    super(JSON.stringify(message))
-    this.name = 'ValidationErrors'
-  }
-}
-
-export { ValidationErrors }
