--- src/lookup-table/utils/validations/types.ts
@@ -1,10 +1,0 @@
-interface IValidator {
-  isValid: boolean
-  errors: string[]
-}
-
-interface ILookupTableNameValidator extends IValidator {
-  validate(): Promise<void> | void
-}
-
-export { IValidator, ILookupTableNameValidator }
